+++ src/store/actions/market/addMarkets.js
@@ -1,5 +1,3 @@
 import { createAction } from 'redux-actions'
 
-export const ADD_MARKET_LIST = 'ADD_MARKET_LIST'
-
-export default createAction(ADD_MARKET_LIST)
+export default createAction('ADD_MARKET_LIST')
