+++ src/store/actions/market/index.js
@@ -1,4 +1,4 @@
-export * from './addMarkets'
-export * from './fetchMarkets'
 export { default as addMarkets } from './addMarkets'
 export { default as fetchMarkets } from './fetchMarkets'
+export { default as updateMarket } from './updateMarket'
+export * from './fetchMarkets'
