+++ src/store/actions/market/updateMarket.js
@@ -1,0 +1,5 @@
+import { createAction } from 'redux-actions'
+
+export const UPDATE_MARKET = 'UPDATE_MARKET'
+
+export default createAction(UPDATE_MARKET)
